Wed Oct 26 17:22:12 2005  Don Busch  <busch_d@ociweb.com>
This is a footprint reduction fix, eliminating a library dependency.
TAO 1.4 introduced a new dependency on the Messaging, Valuetype,
PI, and CodecFactory libraries for the CosNaming_Serv and
CosEvent_Serv libraries.  Turning off CORBA Messaging did not remove
this dependency. Thus, any application that uses CosNaming_Serv or
CosEvent_Serv, but not CORBA Messaging, depends on those libraries
anyway. This change eliminates the CosNaming_Serv and CosEvent_Serv
dependencies on Messaging (and thus PI, CodecFactory, and Valuetype)
when CORBA Messaging and Interceptors are turned off.  It reduces
the footprint in that case by about 600K.


git-svn-id: svn://svn.dre.vanderbilt.edu/DOC/Middleware/trunk@69009 ae88bc3d-4319-0410-8dbf-d08b4c9d3795